Move the DisplayWork-locations serialisation tests to the display lib
